requirements.txt
asgiref==3.5.2
Django==4.1.2
django-bulma==0.8.3
django-filter==22.1
Pillow==9.2.0
pkg_resources==0.0.0
sqlparse==0.4.3
